Unifying genericsom and misrsom
